requirements.txt
aiogram
python-dotenv
sqlalchemy
aiosqlite
gTTS
deep-translator
